pyproject.toml
[tool.poetry]
name = "wealthz"
version = "0.0.1"
description = "Wealth Manager"
authors = ["Daniel Morales <dnlmrls9@gmail.com>"]
repository = "https://github.com/DanielMorales9/wealthz"
documentation = "https://DanielMorales9.github.io/wealthz/"
readme = "README.md"
packages = [
  {include = "wealthz"}
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
polars = "^1.29.0"
google-auth = "^2.40.1"
google-auth-oauthlib = "^1.2.2"
google-api-python-client = "^2.169.0"
duckdb = "^1.2.2"
pydantic = "^2.11.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
pytest-cov = "^6.1.1"
types-pyyaml = "^6.0.12.20250516"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["wealthz"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"



[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov=wealthz --cov-report=term-missing --cov-fail-under=70 --tb=short"
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
